+++ cate/core/opimpl.py
