+++ cate/ds/esa_cci_odp.py
@@ -58,10 +58,10 @@ from owslib.namespaces import Namespaces
 from cate.conf import get_config_value, get_data_stores_path
 from cate.conf.defaults import NETCDF_COMPRESSION_LEVEL
 from cate.core.ds import DATA_STORE_REGISTRY, DataAccessError, DataStore, DataSource, Schema, open_xarray_dataset
+from cate.core.opimpl import subset_spatial_impl, normalize_impl
 from cate.core.types import PolygonLike, TimeLike, TimeRange, TimeRangeLike, VarNamesLike, VarNames
 from cate.ds.local import add_to_data_store_registry, LocalDataSource, LocalDataStore
 from cate.util.monitor import Cancellation, Monitor
-from cate.util.opimpl import subset_spatial_impl, normalize_impl
 
 ESA_CCI_ODP_DATA_STORE_ID = 'esa_cci_odp'
